Changed names of the host package and nodes.
Changed swarm_pose_estimator package name to swarm_centroid_estimator.

Changed centroid_estimator node to centroid_calculator to reflect that the node calculates the centroid with complete information unlike the Kheperas. Plan to create a centroid_estimator node that will run the Khepera functionality on the host machine to increase the swarm size.